Change: Rename ruff module to check
Rename `autohooks.plugins.ruff.ruff` to `autohooks.plugins.ruff.check`
as it does run the ruff linter.
